remove redundant casts

+++ contribs/freight/src/test/java/org/matsim/freight/carriers/CarrierVehicleTypeReaderTest.java
@@ -69,7 +69,7 @@ public class CarrierVehicleTypeReaderTest {
 	@Test
 	void test_whenReadingTypeMedium_itReadsCapacityCorrectly(){
 		VehicleType medium = types.getVehicleTypes().get(Id.create("medium", org.matsim.vehicles.VehicleType.class ) );
-		assertEquals(30., (double) medium.getCapacity().getOther(), Double.MIN_VALUE );
+		assertEquals(30., medium.getCapacity().getOther(), Double.MIN_VALUE );
 	}
 
 	@Test
